# Conflicts:
#	features/merchant/product_detail_common/src/main/java/com/tokopedia/product/detail/common/ProductDetailCommonConstant.kt

+++ features/merchant/atc_variant/src/main/java/com/tkpd/atcvariant/view/viewmodel/AtcVariantViewModel.kt
@@ -365,7 +365,7 @@ class AtcVariantViewModel @Inject constructor(
                shopIdInt: Int,
                categoryName: String,
                userId: String,
-               shippingMinPrice: Int,
+               shippingMinPrice: Double,
                trackerAttributionPdp: String,
                trackerListNamePdp: String,
                isTokoNow: Boolean) {
